Implements in-game notification and error message
effect: eduacate and guide the user ingame.

summary: This provide clear and eye-catching messages to the user when they interact with certain elements. When user presses on certain query buttons, a notification appears explaining associated feature. If nothing happens, a error message will pop up, explaining why.

details:

error_message (new): a function made to draw the error messages on to the screen.

query (new): a function in the class shop_items made to show the information of towers. It pauses the game while  showing the informations.