Patch to the "Stops player from placing towers on enemy path"
Summary: function originally fails to regonise the path between the first checkpoint (0,0) and the second checkpoint. Now I uses an alternative checkpoint.

details:
In pygame.Vector2 the zero vector (0,0) is not recognised as a vector, thus it cannot go through vector manipulation as other checkpoins do.
I now uses (0,1) as the first checkpoint.