Fixed a bug in time calculation when calling dialogues
details:
Problem identifying: originally the  total time paused would  add    (pygametime - time_called - time_level_init) but that does not account for the previously paused times. Subtract that from the (....) and that worked.

Justification:
The reason the pausing time calculation is different for calling dialogues than to all other pausings, is because that the time_called parameter is passed into calling dialogue function,   dialogue_show(time_called: float) and thus the pause beginning time need not to be calculated. Compare it to the method used by other pausings is to call pygame.time before pausing for the pause begin time. This action can cause delay and also be less inefficient when the time_called is already given.